requirements.txt
# Generic
requests == 2.28.1
PyYAML == 5.4.1
Pillow == 9.2.0
# Utils
pytz == 2020.1
urllib3 == 1.26.9
jinja2 == 3.1.1
psutil == 5.9.3
python-dateutil == 2.8.2
# Configurator
expandvars == 0.9.0
# Microservice
apscheduler == 3.9.1
# Microservice - WebService
uvicorn == 0.20.0
Werkzeug == 2.1.2
flask == 2.1.2
quart == 0.17.0
quart-events == 0.4.3
quart_cors == 0.5.0
pluggy == 1.0.0
# Microservice - DBMS
SQLAlchemy == 1.4.36
flask_sqlalchemy == 2.5.1
psycopg2-binary==2.9.3;platform_machine != 'armv7l'
peewee == 3.14.10

# Application
cryptography == 39.0.1
cloudscraper == 1.2.68
beautifulsoup4 == 4.9.3
aniso8601 == 9.0.1
PyMovieDb == 0.0.7
